Rename: 파일 분리
